+++ src/FPath/AppendableFPath.hs
@@ -1,7 +1,3 @@
-{-# LANGUAGE NoImplicitPrelude #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE UnicodeSyntax     #-}
-
 module FPath.AppendableFPath
   ( AppendableFPath( (⫻) ), (</>) )
 where
@@ -30,34 +26,32 @@ import NonEmptyContainers.SeqNEConversions ( IsSeqNonEmpty( seqNE ) )
 
 import FPath.AbsDir   ( AbsDir )
 import FPath.AbsFile  ( AbsFile )
-import FPath.DirType  ( DirTypeC( DirType ) )
 import FPath.RelDir   ( RelDir )
 import FPath.RelFile  ( RelFile )
-import FPath.RelType  ( RelTypeC( RelType ) )
 
 --------------------------------------------------------------------------------
 
 infixl 7 ⫻
 
-class (DirTypeC γ, RelTypeC γ) ⇒ AppendableFPath γ where
-  (⫻) ∷ DirType γ → RelType γ → γ
+class AppendableFPath α β γ | α β → γ where
+  (⫻) ∷ α → β → γ
 
-instance AppendableFPath AbsDir where
+instance AppendableFPath AbsDir RelDir AbsDir where
   d ⫻ f = (d ⊣ seq ⊕ f ⊣ seq) ⊣ re seq
 
-instance AppendableFPath AbsFile where
+instance AppendableFPath AbsDir RelFile AbsFile where
   d ⫻ f = (d ⊣ seq ⪡ f ⊣ seqNE) ⊣ re seqNE
 
-instance AppendableFPath RelDir where
+instance AppendableFPath RelDir RelDir RelDir where
   d ⫻ f = (d ⊣ seq ⊕ f ⊣ seq) ⊣ re seq
 
-instance AppendableFPath RelFile where
+instance AppendableFPath RelDir RelFile RelFile where
   d ⫻ f = (d ⊣ seq ⪡ f ⊣ seqNE) ⊣ re seqNE
 
 infixl 7 </>
 
 -- | non-unicode synonym of `(⫻)`
-(</>) ∷ AppendableFPath γ ⇒ DirType γ → RelType γ → γ
+(</>) ∷ AppendableFPath α β γ ⇒ α → β → γ
 (</>) = (⫻)
 
 -- that's all, folks! ----------------------------------------------------------